+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.4.5
+- `getLayerInfo()` signature changed to require `options`. This caused regressions in down-stream applications. Made `options` default to `{}`.
+- added sanity tests to `feature-service`
+
 ## 0.4.4
 ### Fixed  
-- `getLayerInfo()` signature changed to require `options`. This caused regressions in down-stream applications. Made `options` optional.
+- `getLayerInfo()` signature changed to require `options`. This caused regressions in down-stream applications. Made `options` default to null.
 
 ## 0.4.3
 ### Fixed